refactor: move the store and the store hook out to their own module
